// src/blog/data-source/legacy-strapi-types.ts
export type Article = {
  id: number;
  attributes: {
    title: string;
    description: string;
    content: string;
    slug: string;
    createdAt: string;
    updatedAt: string;
    publishedAt: string;
    postedTime: string;
    displayTime: Date;
    pin?: boolean | null;
    publicationState?: PublicationState;
    cover: {
      data: {
        id: string;
        attributes: ShareMedia;
      };
    };
    category: {
      data: {
        id: number;
        attributes: {
          name: string;
          slug: string;
          description: string;
          createdAt: string;
          updatedAt: string;
        };
      };
    };
    author: {
      data: {
        id: number;
        attributes: {
          name: string;
          email: string;
        };
      };
    };
    blocks: Block[];
  };
};

export type Block = {
  id: number;
  __component: string;
  title?: string;
  body?: string;
  url?: string;
  html?: string;
  ext?: string;
  caption?: string;
  file?: {
    data: {
      id: string;
      attributes: ShareMedia;
    };
  };
};

export type Categorg = {
  id: number;
  attributes: {
    name: string;
    slug: string;
    description: string;
    createdAt: string;
    updatedAt: string;
  };
};

export type TPagination = {
  page: number;
  pageSize: number;
  pageCount: number;
  total: number;
};

export type TFile = {
  id: number;
  url: string;
  name: string;
  width: number;
  height: number;
};

export type ShareMedia = {
  name: string;
  url: string;
  alternativeText: string;
  caption: string;
  formats: {
    small: {
      ext: string;
      url: string;
      name: string;
    };
    medium: {
      ext: string;
      url: string;
      name: string;
    };
    thumbnail: {
      ext: string;
      url: string;
      name: string;
    };
  };
  related?: any[];
  ext: string;
  mime: string;
};

export type Pagination = {
  page?: number;
  pageSize?: number;
  withCount?: boolean;
};

export type PublicationState = "live" | "preview";

export enum PublicationStateEnum {
  // 获取已发布的文章，用于生产环境
  LIVE = "live",
  // 获取包含未发布和已发布的文章，用于开发环境
  PREVIEW = "preview",
}

export type ResponstList<T> = {
  data: T;
  meta: { pagination: TPagination };
};